Fixing DagBag tests
