+++ tests/models/__init__.py
@@ -16,11 +16,4 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-
-# flake8: noqa
-
-from __future__ import absolute_import
-
-from .api import *
-from .core import *
-from .test_models import *
+#
